Move packet nonce field to packet instead of packet header
