+++ crates/ibc/ibc-components/src/impls/handlers/incoming/packet/no_replay.rs
@@ -6,8 +6,8 @@ use hermes_chain_type_components::traits::types::commitment_proof::HasCommitment
 use hermes_chain_type_components::traits::types::ibc::channel_id::HasChannelIdType;
 
 use crate::traits::fields::packet::header::channel_id::HasPacketChannelIds;
-use crate::traits::fields::packet::header::nonce::HasPacketNonce;
 use crate::traits::fields::packet::packet::header::HasPacketHeader;
+use crate::traits::fields::packet::packet::nonce::HasPacketNonce;
 use crate::traits::handlers::incoming::packet::IncomingPacketHandler;
 use crate::traits::queries::recv_packet_commitment::CanQueryHasPacketReceived;
 use crate::traits::types::packet::nonce::HasPacketNonceType;
@@ -44,7 +44,7 @@ where
         send_proof: &Counterparty::CommitmentProof,
     ) -> Result<(), Chain::Error> {
         let packet_header = Counterparty::packet_header(packet);
-        let nonce = Counterparty::packet_nonce(packet_header);
+        let nonce = Counterparty::packet_nonce(packet);
         let src_channel_id = Counterparty::packet_src_channel_id(packet_header);
         let dst_channel_id = Counterparty::packet_dst_channel_id(packet_header);
