+++ crates/ibc/ibc-components/src/traits/fields/packet/header/mod.rs
@@ -1,3 +1,2 @@
 pub mod channel_id;
-pub mod nonce;
 pub mod timeout;
