+++ crates/ibc/ibc-components/src/traits/fields/packet/packet/mod.rs
@@ -1,2 +1,3 @@
 pub mod header;
+pub mod nonce;
 pub mod payloads;
