+++ crates/ibc/ibc-components/src/types/packet.rs
@@ -4,15 +4,19 @@ use alloc::vec::Vec;
 use cgp::prelude::*;
 
 use crate::traits::types::packet::header::HasPacketHeaderType;
+use crate::traits::types::packet::nonce::HasPacketNonceType;
 use crate::traits::types::packet::packet::ProvidePacketType;
 use crate::traits::types::payload::payload::HasPayloadType;
 
 #[derive(HasField)]
 pub struct IbcPacket<Chain, Counterparty>
 where
-    Chain: HasPacketHeaderType<Counterparty> + HasPayloadType<Counterparty>,
+    Chain: HasPacketHeaderType<Counterparty>
+        + HasPacketNonceType<Counterparty>
+        + HasPayloadType<Counterparty>,
 {
     pub header: Chain::PacketHeader,
+    pub nonce: Chain::PacketNonce,
     pub payloads: Vec<Chain::Payload>,
 }
 
@@ -20,7 +24,9 @@ pub struct UseIbcPacket<App>(pub PhantomData<App>);
 
 impl<Chain, Counterparty, App> ProvidePacketType<Chain, Counterparty> for UseIbcPacket<App>
 where
-    Chain: HasPacketHeaderType<Counterparty> + HasPayloadType<Counterparty>,
+    Chain: HasPacketHeaderType<Counterparty>
+        + HasPacketNonceType<Counterparty>
+        + HasPayloadType<Counterparty>,
     Counterparty: Async,
     App: Async,
 {
@@ -30,11 +36,13 @@ where
 impl<Chain, Counterparty> Clone for IbcPacket<Chain, Counterparty>
 where
     Chain: HasPacketHeaderType<Counterparty, PacketHeader: Clone>
+        + HasPacketNonceType<Counterparty, PacketNonce: Clone>
         + HasPayloadType<Counterparty, Payload: Clone>,
 {
     fn clone(&self) -> Self {
         Self {
             header: self.header.clone(),
+            nonce: self.nonce.clone(),
             payloads: self.payloads.clone(),
         }
     }
