+++ crates/ibc/ibc-components/src/types/packet_header.rs
@@ -2,18 +2,16 @@ use cgp::prelude::*;
 use hermes_chain_type_components::traits::types::ibc::channel_id::HasChannelIdType;
 
 use crate::traits::types::packet::header::ProvidePacketHeaderType;
-use crate::traits::types::packet::nonce::HasPacketNonceType;
 use crate::traits::types::packet::timeout::HasPacketTimeoutType;
 
 #[derive(HasField)]
 pub struct IbcPacketHeader<Chain, Counterparty>
 where
-    Chain: HasChannelIdType<Counterparty> + HasPacketNonceType<Counterparty>,
+    Chain: HasChannelIdType<Counterparty>,
     Counterparty: HasChannelIdType<Chain> + HasPacketTimeoutType<Chain>,
 {
     pub src_channel_id: Chain::ChannelId,
     pub dst_channel_id: Counterparty::ChannelId,
-    pub nonce: Chain::PacketNonce,
     pub timeout: Counterparty::PacketTimeout,
 }
 
@@ -21,7 +19,7 @@ pub struct UseIbcPacketHeader;
 
 impl<Chain, Counterparty> ProvidePacketHeaderType<Chain, Counterparty> for UseIbcPacketHeader
 where
-    Chain: HasChannelIdType<Counterparty> + HasPacketNonceType<Counterparty>,
+    Chain: HasChannelIdType<Counterparty>,
     Counterparty: HasChannelIdType<Chain> + HasPacketTimeoutType<Chain>,
 {
     type PacketHeader = IbcPacketHeader<Chain, Counterparty>;
@@ -29,8 +27,7 @@ where
 
 impl<Chain, Counterparty> Clone for IbcPacketHeader<Chain, Counterparty>
 where
-    Chain: HasChannelIdType<Counterparty, ChannelId: Clone>
-        + HasPacketNonceType<Counterparty, PacketNonce: Clone>,
+    Chain: HasChannelIdType<Counterparty, ChannelId: Clone>,
     Counterparty: HasChannelIdType<Chain, ChannelId: Clone>
         + HasPacketTimeoutType<Chain, PacketTimeout: Clone>,
 {
@@ -38,7 +35,6 @@ where
         Self {
             src_channel_id: self.src_channel_id.clone(),
             dst_channel_id: self.dst_channel_id.clone(),
-            nonce: self.nonce.clone(),
             timeout: self.timeout.clone(),
         }
     }
