+++ crates/ibc/ibc-mock-chain/src/components/chain.rs
@@ -10,8 +10,8 @@ use hermes_chain_type_components::traits::types::height::HeightTypeComponent;
 use hermes_chain_type_components::traits::types::ibc::channel_id::ChannelIdTypeComponent;
 use hermes_ibc_components::traits::fields::message::app_id::IbcMessageAppIdGetterComponent;
 use hermes_ibc_components::traits::fields::packet::header::channel_id::PacketChannelIdGetterComponent;
-use hermes_ibc_components::traits::fields::packet::header::nonce::PacketNonceGetterComponent;
 use hermes_ibc_components::traits::fields::packet::header::timeout::PacketTimeoutGetterComponent;
+use hermes_ibc_components::traits::fields::packet::packet::nonce::PacketNonceGetterComponent;
 use hermes_ibc_components::traits::fields::packet::packet::payloads::PacketPayloadsGetterComponent;
 use hermes_ibc_components::traits::fields::payload::app_id::PayloadAppIdGetterComponent;
 use hermes_ibc_components::traits::types::app_id::AppIdTypeComponent;
