+++ crates/ibc/ibc-mock-chain/src/contexts/chain.rs
@@ -12,8 +12,8 @@ use hermes_chain_type_components::traits::types::height::HasHeightType;
 use hermes_chain_type_components::traits::types::ibc::channel_id::HasChannelIdType;
 use hermes_ibc_components::traits::fields::message::app_id::HasIbcMessageAppIds;
 use hermes_ibc_components::traits::fields::packet::header::channel_id::HasPacketChannelIds;
-use hermes_ibc_components::traits::fields::packet::header::nonce::HasPacketNonce;
 use hermes_ibc_components::traits::fields::packet::header::timeout::HasPacketTimeout;
+use hermes_ibc_components::traits::fields::packet::packet::nonce::HasPacketNonce;
 use hermes_ibc_components::traits::fields::packet::packet::payloads::HasPacketPayloads;
 use hermes_ibc_components::traits::fields::payload::app_id::HasPayloadAppIds;
 use hermes_ibc_components::traits::types::app_id::HasAppIdType;
